Moved LoginResult to models folder.
