+++ lib/services/i_auth_service.dart
@@ -1,13 +1,4 @@
-class LoginResult {
-  final bool success;
-  final String id;
-
-  LoginResult.withId(this.id) : success = true;
-
-  LoginResult.failed()
-      : success = false,
-        id = null;
-}
+import 'package:chatroom/models/login_result.dart';
 
 abstract class IAuthService {
   Future<LoginResult> logIn();
